# plugin/cache: unroll minTTL loop